requirements.txt
Package                 Version
----------------------- ---------
boto3                   1.21.21
botocore                1.24.21
certifi                 2021.10.8
charset-normalizer      2.0.12
docopt                  0.6.2
gpytorch                1.5.1
idna                    3.3
jmespath                1.0.0
joblib                  1.1.0
mkl-fft                 1.3.1
mkl-random              1.2.2
mkl-service             2.4.0
numpy                   1.21.2
olefile                 0.46
Pillow                  8.4.0
pip                     21.2.4
pipreqs                 0.4.11
python-dateutil         2.8.2
pytorch-pretrained-bert 0.6.2
regex                   2022.3.15
requests                2.27.1
s3transfer              0.5.2
scikit-learn            1.0.1
scipy                   1.7.3
setuptools              58.0.4
six                     1.16.0
threadpoolctl           3.0.0
torch                   1.10.0
torchaudio              0.10.0
torchvision             0.11.1
tqdm                    4.63.0
typing-extensions       3.10.0.2
urllib3                 1.26.9
wheel                   0.37.0
yarg                    0.1.9
